+++ src/frontend/src/features/setting/components/setting-profile.tsx
@@ -1,4 +1,6 @@
-import { fetchProfileImage, uploadProfileImage } from '@/lib/setting/setting-profile'
+import { fetchProfileImage } from '@/lib/setting/get-profileImage'
+import { uploadProfileImage } from '@/lib/setting/post-profileImage'
+
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 import { type ChangeEvent, useState } from 'react'
 import { FiCamera } from 'react-icons/fi'
